fix: register model to xpert module
