+++ apps/cloud/src/app/xpert/index.ts
@@ -1,3 +1,4 @@
 export * from './conversation/conversation.component'
 export * from './chat.service'
 export * from './chat-input/chat-input.component'
+export * from './home.service'
